🔧 fix : delete post 시 포스트 작성자가 해당 멤버인지 확인

+++ src/main/java/web/fridge/domain/post/controller/PostController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import web.fridge.domain.member.annotation.AuthMember;
 import web.fridge.domain.member.entity.Member;
+import web.fridge.domain.member.service.MemberService;
 import web.fridge.domain.post.dto.PostDetails;
 import web.fridge.domain.post.entity.Post;
 import web.fridge.domain.post.service.*;
@@ -24,6 +25,7 @@ public class PostController {
     private final PostFindService postFindService;
     private final PostRegisterService postRegisterService;
     private final PostRemoveService postRemoveService;
+    private final MemberService memberService;
 
     @PostMapping
     public ResponseEntity<Post> postAdd(
@@ -35,15 +37,23 @@ public class PostController {
         return new ResponseEntity<>(post, HttpStatus.OK);
     }
 
+    //포스트 삭제 시 관련 거래도 삭제 되게 할건지? 다른 유저의 거래 기록 확인이 남아있어야하니까
+    //채팅 추가 시 관련 채팅방 삭제도 해야함
     @DeleteMapping("/{postId}")
     public ResponseEntity<String> postRemove(
             @AuthMember Member member,
             @PathVariable(value = "postId") Long postId){
         log.info("[PostController][postRemove]:{}",postId);
-        postRemoveService.removePost(postId);
-        return new ResponseEntity<>("delete post",HttpStatus.OK);
+        Boolean isPostOwner = memberService.isPostOwner(member,postId);
+        if(isPostOwner == Boolean.TRUE){
+            postRemoveService.removePost(postId);
+            return new ResponseEntity<>("delete post",HttpStatus.OK);
+        }
+        else return new ResponseEntity<>("your not a post owner",HttpStatus.OK);
     }
-    
+
+    //거래 완료된 포스트도 가져올 건지?
+    //post 등록 시 address 등록 형식 문제가 있음 (자유롭게 String으로 받으면 멤버의 Region과 비교하기 어려움)
     @GetMapping
     public ResponseEntity<List<Post>> postByRegionList(
             @AuthMember Member member){
@@ -56,7 +66,7 @@ public class PostController {
     public ResponseEntity<Post> postByPostID(
             @AuthMember Member member,
             @PathVariable(value = "postId") Long postId){
-        Post post = postFindService.findPostByPostID(member,postId);
+        Post post = postFindService.findPostByPostID(postId);
         return new ResponseEntity<>(post,HttpStatus.OK);
     }
 }
